+++ frontend/src/App.tsx
@@ -1,24 +1,19 @@
-
-import React from 'react'
+import './reset.css'
 import { Routes, Route } from 'react-router-dom'
-
-import MainPage from './pages/MainPage'
-import Header from './components/Header'
-import BottomNavigation from './components/BottomNavigation'
-import Login from './pages/Login'
-
+import LoginPage from './pages/LoginPage'
 
 function App() {
-  return (
-    <>
-      <Header />
-      <Routes>
-        <Route path="/" element={<MainPage />} />
-        <Route path="/login" element={<Login />} />
-      </Routes>
-      <BottomNavigation />
-    </>
+  // const [count, setCount] = useState(0)
+  // useEffect(() => {
+  //   fetch('/api/test')
+  //     .then(response => response.json())
+  //     .then(json => setMessage(json.SUCCESS_TEXT))
+  // }, [])
 
+  return (
+    <Routes>
+      <Route path="/" element={<LoginPage />} />
+    </Routes>
   )
 }
